service/app_ami/requirements.txt
asterisk-ami==0.1.7
pika==1.3.2
python-dotenv==1.0.1
